Spanning Trees

Erdos.count_spanning_trees — Method

count_spanning_trees(g::AGraph)

Returns the number of spanning trees of g, computed through Kirchhoff's theorem. The return type is a float, since the number can be very large.

Erdos.minimum_spanning_tree — Method

minimum_spanning_tree{T<:Real}(
    g, distmx::AbstractMatrix{T} = weights(g)
)

Performs Kruskal's algorithm on a connected, undirected graph g, having adjacency matrix distmx, and computes minimum spanning tree. Returns a Vector{KruskalHeapEntry}, that contains the containing edges and its weights.